refactor: mv {OverlayDB,validate_json} from .util to .api.db.mongo
rm redundant {UpdateStatement,DeleteStatement} defs in .util and have OverlayDB source defs from .api.models.query

towards #1007
