+++ nmdc_runtime/api/endpoints/metadata.py
@@ -15,7 +15,7 @@ from gridfs import GridFS, NoFile
 from jsonschema import Draft7Validator
 from nmdc_runtime.api.core.metadata import _validate_changesheet, df_from_sheet_in
 from nmdc_runtime.api.core.util import API_SITE_CLIENT_ID
-from nmdc_runtime.api.db.mongo import get_mongo_db
+from nmdc_runtime.api.db.mongo import get_mongo_db, validate_json
 from nmdc_runtime.api.endpoints.util import (
     _claim_job,
     _request_dagster_run,
@@ -30,7 +30,6 @@ from nmdc_runtime.api.models.user import User, get_current_active_user
 from nmdc_runtime.site.repository import repo, run_config_frozen__normal_env
 from nmdc_runtime.util import (
     unfreeze,
-    validate_json,
     specialize_activity_set_docs,
 )
 from nmdc_runtime.util import get_nmdc_jsonschema_dict
