+++ nmdc_runtime/api/models/query.py
@@ -1,4 +1,3 @@
-import datetime
 import json
 import logging
 from typing import Optional, Any, Dict, List, Union, TypedDict
@@ -12,16 +11,12 @@ from pydantic import (
     PositiveInt,
     NonNegativeInt,
     field_validator,
-    ConfigDict,
     WrapSerializer,
 )
-from pymongo.database import Database as MongoDatabase
 from toolz import assoc, assoc_in
 from typing_extensions import Annotated
 
-from nmdc_runtime.api.core.idgen import generate_one_id
-from nmdc_runtime.api.core.util import now, pick
-from nmdc_runtime.api.db.mongo import get_mongo_db
+from nmdc_runtime.api.core.util import pick
 
 
 def bson_to_json(doc: Any, handler) -> dict:
@@ -156,6 +151,7 @@ class CursorYieldingCommandResponse(CommandResponse):
 
 class DeleteStatement(BaseModel):
     q: Document
+    # `limit` is required: https://www.mongodb.com/docs/manual/reference/command/delete/#std-label-deletes-array-limit
     limit: OneOrZero
     hint: Optional[Dict[str, OneOrMinusOne]] = None
 
@@ -245,6 +241,3 @@ def command_response_for(type_):
         UpdateCommand: UpdateCommandResponse,
     }
     return d.get(type_)
-
-
-_mdb = get_mongo_db()
